Moved API framework sources from lib to src
This separates the source relocation from the later TypeScript and ESM conversion so file history remains readable.
